test(compose): split retry validation coverage
